__app__/requirements.txt
# azure # with out this get "Exception: ImportError: cannot import name 'BlockBlobService' from 'azure.storage.blob'" even with "azure-storage-blob" below
azure-storage-blob>=12.5.0
azure-functions
azure-servicebus>=7.0.0b7
aiohttp
aiodns
cchardet